Move replay-strategy.ts
This commit moves the replay-strategy.ts module to the FutureProcessor's
helpers folder, as it's only meant to be used there.
